refactor: move all classes to their own  folder
