+++ src/cliente/Cliente.hpp
@@ -4,7 +4,7 @@
 #include <string>
 using namespace std;
 
-class Cliente{
+class Cliente {
 
 	public:
     Cliente();
