+++ src/encomenda/Encomenda.hpp
@@ -1,7 +1,7 @@
 #ifndef Encomenda_H
 #define Encomenda_H
 
-#include "Cliente.hpp"
+#include "../cliente/Cliente.hpp"
 using namespace std;
 
 class Encomenda{
